% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linput3.0.5.R
\name{linput}
\alias{linput}
\title{This package was built by Yunhan DU(1270950448@qq.com) & Chen Liang(liangchen139@163.com),from Southwestern University of Finance and Economics,China,Sichuan Province===Beijing Forestry University,Beijing.
it can solve the problem that some economic data has NAs.}
\usage{
linput(d, p, y, x, m)
}
\arguments{
\item{d}{an dataframe object for which a panel data has some NA values.}

\item{p}{The variable which indicate the individual.}

\item{y}{The variable which indicate time series variable of an individual.}

\item{x}{The variable which indicate the variable has NA values and need to be input.}

\item{m}{m = 0/1,indicating only 'interpolation method'/both 'extrapolation and interpolation method' would be chosen.}
}
\description{
Details:
linput is a generic function used to proceed with the NA values of panel data.
The function invokes particular methods (only interpolation or both extrapolation and interpolation) which depends on the value of the fifth argument m.
1.This package was built to solve the problem that some economic data has NAs by using linear input(method:interpolation) formula and average growth rate per annum(method:extrapolation) to input NAs.
2.For this package named linput, it can input time series variable of an individual with at least two values, if the variable has no values or only one value, Error information would be presented.
3.The time series variable of individuals with input would be saved in a temporary R dataframe named test_data2, The input variable would be named as testemploy of test_data2,and the input value also would be saved in a new R value named estemploy_i, i depends on the number of times you run this package linput. Every time you run this package to proceed your data, the input result would be saved in a new R value named testemployi(i=1,2,...) automaticly.
}
\examples{
linput(user_data,user_data$province,user_data$year,user_data$income,m = 1)
linput(user_data,user_data$province,user_data$year,user_data$income,m = 0)
}
